update todo fragments - refactoring
