+++ app/src/main/java/com/diego/discoteca/data/source/paging/DatabasePagingSource.kt
@@ -5,9 +5,9 @@ import androidx.paging.PagingState
 import com.diego.discoteca.data.domain.Disc
 import com.diego.discoteca.data.model.DiscDb
 import com.diego.discoteca.database.DiscDatabaseDao
-import com.diego.discoteca.database.asDomainModel
 import com.diego.discoteca.util.AddBy
 import com.diego.discoteca.util.Constants.DATABASE_STARTING_PAGE_INDEX
+import com.diego.discoteca.util.asDomainModel
 import com.diego.discoteca.util.stringNormalizeDatabase
 
 class DatabasePagingSourceBarcode(
