+++ app/src/main/java/com/diego/discoteca/ui/discDetail/DiscDetailFragment.kt
@@ -2,10 +2,16 @@ package com.diego.discoteca.ui.discDetail
 
 import android.graphics.Color
 import android.os.Bundle
-import android.view.*
-import androidx.databinding.DataBindingUtil
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
+import android.view.View
+import androidx.core.view.MenuHost
+import androidx.core.view.MenuProvider
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.Lifecycle
+import androidx.navigation.fragment.navArgs
 import com.diego.discoteca.DiscotecaApplication
 import com.diego.discoteca.R
 import com.diego.discoteca.databinding.FragmentDiscDetailBinding
@@ -13,15 +19,13 @@ import com.diego.discoteca.ui.activity.MainActivity
 import com.diego.discoteca.util.themeColor
 import com.google.android.material.transition.MaterialContainerTransform
 
-//TODO : update deprecation
-@Suppress("DEPRECATION")
-class DiscDetailFragment : Fragment() {
+class DiscDetailFragment : Fragment(R.layout.fragment_disc_detail), MenuProvider {
 
+    private val args: DiscDetailFragmentArgs by navArgs()
     private val mDiscDetailViewModel: DiscDetailViewModel by viewModels {
-        val arguments = DiscDetailFragmentArgs.fromBundle(requireArguments())
         DiscDetailViewModelFactory(
             repository = (requireContext().applicationContext as DiscotecaApplication).discsRepository,
-            discId = arguments.discId
+            discId = args.discId
         )
     }
 
@@ -35,16 +39,9 @@ class DiscDetailFragment : Fragment() {
         }
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View {
-        val binding: FragmentDiscDetailBinding = DataBindingUtil.inflate(
-            inflater,
-            R.layout.fragment_disc_detail,
-            container,
-            false
-        )
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val binding = FragmentDiscDetailBinding.bind(view)
         binding.apply {
             lifecycleOwner = viewLifecycleOwner
             discDetailViewModel = mDiscDetailViewModel
@@ -57,13 +54,23 @@ class DiscDetailFragment : Fragment() {
             }
         }
 
-        setHasOptionsMenu(true)
-        return binding.root
+        setupMenu()
     }
 
-    @Deprecated("Deprecated in Java")
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
+    override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
         menu.clear()
+    }
+
+    override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
+        return false
+    }
+
+    private fun setupMenu() {
+        val menuHost: MenuHost = requireActivity()
+        menuHost.addMenuProvider(
+            this,
+            viewLifecycleOwner,
+            Lifecycle.State.STARTED
+        )
     }
 }
