+++ app/src/main/java/com/diego/discoteca/ui/discPresentDetail/DiscPresentDetailFragment.kt
@@ -2,10 +2,15 @@ package com.diego.discoteca.ui.discPresentDetail
 
 import android.graphics.Color
 import android.os.Bundle
-import android.view.*
-import androidx.databinding.DataBindingUtil
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
+import android.view.View
+import androidx.core.view.MenuProvider
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.Lifecycle
+import androidx.navigation.fragment.navArgs
 import com.diego.discoteca.DiscotecaApplication
 import com.diego.discoteca.R
 import com.diego.discoteca.data.model.DiscPresent
@@ -16,15 +21,13 @@ import com.diego.discoteca.util.themeColor
 import com.google.android.material.transition.MaterialContainerTransform
 import com.google.android.material.transition.MaterialSharedAxis
 
-//TODO : update deprecation
-@Suppress("DEPRECATION")
-class DiscPresentDetailFragment : Fragment() {
+class DiscPresentDetailFragment : Fragment(R.layout.fragment_disc_present_detail), MenuProvider {
 
+    private val args: DiscPresentDetailFragmentArgs by navArgs()
     private val mDiscPresentDetailViewModel: DiscPresentDetailViewModel by viewModels {
-        val arguments = DiscPresentDetailFragmentArgs.fromBundle(requireArguments())
         DiscPresentDetailViewModelFactory(
             repository = (requireContext().applicationContext as DiscotecaApplication).discsRepository,
-            discPresent = arguments.discPresent
+            discPresent = args.discPresent
         )
     }
 
@@ -38,16 +41,9 @@ class DiscPresentDetailFragment : Fragment() {
         }
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View {
-        val binding: FragmentDiscPresentDetailBinding = DataBindingUtil.inflate(
-            inflater,
-            R.layout.fragment_disc_present_detail,
-            container,
-            false
-        )
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val binding = FragmentDiscPresentDetailBinding.bind(view)
         binding.apply {
             lifecycleOwner = viewLifecycleOwner
             discPresentDetailViewModel = mDiscPresentDetailViewModel
@@ -69,14 +65,24 @@ class DiscPresentDetailFragment : Fragment() {
             }
         }
 
-        setHasOptionsMenu(true)
-        return binding.root
+        setupMenu()
     }
 
-    @Deprecated("Deprecated in Java")
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
+    override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
         menu.clear()
+    }
+
+    override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
+        return false
+    }
+
+    private fun setupMenu() {
+        val menuHost = requireActivity()
+        menuHost.addMenuProvider(
+            this,
+            viewLifecycleOwner,
+            Lifecycle.State.STARTED
+        )
     }
 
     private fun navigatePopStack() {
