+++ app/src/main/java/com/diego/discoteca/ui/information/InformationFragment.kt
@@ -1,10 +1,14 @@
 package com.diego.discoteca.ui.information
 
 import android.os.Bundle
-import android.view.*
-import androidx.databinding.DataBindingUtil
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
+import android.view.View
+import androidx.core.view.MenuProvider
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import androidx.lifecycle.Lifecycle
 import com.diego.discoteca.DiscotecaApplication
 import com.diego.discoteca.R
 import com.diego.discoteca.adapter.CountFormatMediaAdapter
@@ -13,9 +17,7 @@ import com.diego.discoteca.ui.activity.MainActivity
 import com.diego.discoteca.util.Destination
 import com.google.android.material.transition.MaterialFadeThrough
 
-//TODO : update deprecation
-@Suppress("DEPRECATION")
-class InformationFragment : Fragment() {
+class InformationFragment : Fragment(R.layout.fragment_information), MenuProvider {
 
     private val mInformationViewModel: InformationViewModel by viewModels {
         InformationViewModelFactory(
@@ -30,18 +32,9 @@ class InformationFragment : Fragment() {
         }
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        val binding: FragmentInformationBinding = DataBindingUtil.inflate(
-            inflater,
-            R.layout.fragment_information,
-            container,
-            false
-        )
-
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val binding = FragmentInformationBinding.bind(view)
         val adapterCountFormatMedia = CountFormatMediaAdapter()
         binding.apply {
             lifecycleOwner = viewLifecycleOwner
@@ -63,14 +56,24 @@ class InformationFragment : Fragment() {
             }
         }
 
-        setHasOptionsMenu(true)
-        return binding.root
+        setupMenu()
     }
 
-    @Deprecated("Deprecated in Java")
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
+    override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
         menu.clear()
+    }
+
+    override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
+        return false
+    }
+
+    private fun setupMenu() {
+        val menuHost = requireActivity()
+        menuHost.addMenuProvider(
+            this,
+            viewLifecycleOwner,
+            Lifecycle.State.STARTED
+        )
     }
 
     private fun goToScanBarcodeFragment() {
