+++ build.gradle
@@ -36,7 +36,7 @@ ext {
     ktxVersion = '1.9.0'
     lifecycleExtensionsVersion = '2.2.0'
     lifecycleVersion = '2.5.1'
-    materialVersion = '1.6.1'
+    materialVersion = '1.7.0'
     mlkitVersion = '17.0.2'
     moshiVersion = '1.13.0'
     moshiConverterVersion = '2.9.0'
